[MGMIDN-1479]: fix api for score and learning time per month.

+++ server/src/main/java/com/mgmtp/cfu/dto/userdto/UserScore.java
@@ -15,6 +15,6 @@ public class UserScore {
     private long score;
     private String year;
     private List<String> months;
-    private List<Double> scores;
-    private List<Double> learningTime;
+    private List<Integer> scores;
+    private List<Integer> learningTime;
 }
